requirements.txt
streamlit
mysql-connector-python
pandas
scikit-learn
pdfplumber
plotly
qrcode
Pillow
fpdf2
bcrypt
